integrations/requirements.txt
requests
python-dotenv
dataclasses
pymongo